Improve style and move out of io module
